backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-multipart==0.0.20
websockets==14.1
google-generativeai==0.8.3
langgraph==0.2.60
langchain-core==0.3.28
langchain-google-genai==2.0.7
pydantic-settings==2.7.0
chromadb==0.5.23
Pillow==11.0.0
pypdf2==3.0.1
python-dotenv==1.0.1
greenlet==3.1.1
